+++ tests/testthat/testSimulateReplicates.R
@@ -7,7 +7,7 @@ seed <- 1
 source(paste0("", "testUtils.R"))
 
 test_that(getTestName("VPC on CP, using predicate"), {
-  if (skipLongTest) return(TRUE)
+  if (skipLongTests()) return(TRUE)
   model <- model_suite$testing$other$my_model1
   model <- model %>% disable(c("VARCOV_OMEGA", "VARCOV_SIGMA"))
   regFilename <- "full_uncertainty"
@@ -24,7 +24,7 @@ test_that(getTestName("VPC on CP, using predicate"), {
 })
 
 test_that(getTestName("VPC on both CP and Y, using function"), {
-  if (skipLongTest) return(TRUE)
+  if (skipLongTests()) return(TRUE)
   model <- model_suite$testing$other$my_model1
   model <- model %>% disable(c("VARCOV_OMEGA", "VARCOV_SIGMA"))
   regFilename <- "full_uncertainty"
@@ -48,9 +48,8 @@ test_that(getTestName("VPC on both CP and Y, using function"), {
 })
 
 test_that(getTestName("Study replication also works with scenarios"), {
-  if (skipLongTest) return(TRUE)
-  if (skipVdiffrTest) return(TRUE)
-  
+  if (skipLongTests()) return(TRUE)
+
   model <- model_suite$testing$nonmem$advan2_trans1
   ds <- Dataset(10) %>%
     add(Bolus(time=0, amount=1000)) %>%
@@ -59,17 +58,19 @@ test_that(getTestName("Study replication also works with scenarios"), {
   scenarios <- Scenarios() %>%
     add(Scenario(name="Base model")) %>%
     add(Scenario(name="Increased KA", model=~.x %>% replace(Theta(name="KA", value=3)))) # 3 instead of 1
-  
+
   # Outfun executed at the level of the scenario (backwards compatibility)
   simulation <- expression(simulate(model=model, dataset=ds, dest=destEngine, replicates=5,
                                     outfun=~PI(.x, output="CP"), seed=seed, scenarios=scenarios))
   test <- expression(
     expect_true(all(c("replicate", "TIME", "metric", "value", "SCENARIO") %in% colnames(results))),
     expect_true(all(results$SCENARIO %>% unique()==c("Base model", "Increased KA"))),
-    vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    if (!skipVdiffrTests()) {
+      vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    }
   )
   campsisTest(simulation, test, env=environment())
-  
+
   # Outfun executed at the level of the replicate (possible since Campsis v1.5.3)
   simulation <- expression(simulate(model=model, dataset=ds, dest=destEngine, replicates=5,
                                     outfun=Outfun(~PI(.x, output="CP", scenarios="SCENARIO"), level="replicate"),
@@ -77,10 +78,12 @@ test_that(getTestName("Study replication also works with scenarios"), {
   test <- expression(
     expect_true(all(c("replicate", "TIME", "metric", "value", "SCENARIO") %in% colnames(results))),
     expect_true(all(results$SCENARIO %>% unique()==c("Base model", "Increased KA"))),
-    vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    if (!skipVdiffrTests()) {
+      vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    }
   )
   campsisTest(simulation, test, env=environment())
-  
+
   # Alternatively, function and arguments may also be passed (possible since Campsis v1.5.3)
   # This is particularly useful when parallelisation on replicates is enabled since
   # the function written in a lambda will not be detected as part of the environment by the 'future' package
@@ -90,13 +93,15 @@ test_that(getTestName("Study replication also works with scenarios"), {
   test <- expression(
     expect_true(all(c("replicate", "TIME", "metric", "value", "SCENARIO") %in% colnames(results))),
     expect_true(all(results$SCENARIO %>% unique()==c("Base model", "Increased KA"))),
-    vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    if (!skipVdiffrTests()) {
+      vdiffr::expect_doppelganger("VPC / specified outfun", vpcPlot(results, scenarios="SCENARIO") + facet_wrap(~SCENARIO))
+    }
   )
   campsisTest(simulation, test, env=environment())
 })
 
 test_that(getTestName("Try/catch works as expected if one replicate fails"), {
-  if (skipLongTest) return(TRUE)
+  if (skipLongTests()) return(TRUE)
   model <- model_suite$testing$nonmem$advan2_trans2
 
   # Add high uncertainty on THETA_KA (variance of 1)
@@ -132,7 +137,7 @@ test_that(getTestName("Try/catch works as expected if one replicate fails"), {
 })
 
 test_that(getTestName("Replicates can be simulated in parallel"), {
-  if (skipLongTest) return(TRUE)
+  if (skipLongTests()) return(TRUE)
   # progressr::handlers(global=TRUE)
   # progressr::handlers(campsis_handler())
   regFilename <- "replicates_in_parallel"