+++ NEWS.md
@@ -1,3 +1,11 @@
+# campsis 1.5.5
+
+* Arm label mapping must first verify the ARM column exists #155
+* Code coverage does not appear on codecov.io anymore #161
+* Optionally skip tests using R options #162
+* Do not build vignettes on CRAN #163
+* Update pkgdown documentation #164
+
 # campsis 1.5.4
 
 * Add 'label' argument to 'Dataset' constructor #151